Models placed their own classes
