+++ Models/AuthenticationModels/UserLoginRequest.cs
@@ -1,0 +1,8 @@
+﻿namespace LethalCompany_Backend.Models.AuthenticationModels;
+
+public class UserLoginRequest
+{
+    public string Email { get; set; }
+
+    public string Password { get; set; }
+}
