--- Models/InterestModels.cs
@@ -1,9 +1,0 @@
-﻿namespace LethalCompany_Backend.Models
-{
-    public class InterestViewModel
-    {
-        public long Id { get; set; }
-        public string Name { get; set; }
-        public bool IsChecked { get; set; }
-    }
-}
