+++ Models/InterestModels/InterestViewModel.cs
@@ -1,0 +1,8 @@
+﻿namespace LethalCompany_Backend.Models.InterestModels;
+
+public class InterestViewModel
+{
+    public long Id { get; set; }
+    public string Name { get; set; }
+    public bool IsChecked { get; set; }
+}
