+++ Models/MailSenderModels/MailConfig.cs
@@ -1,0 +1,10 @@
+﻿namespace LethalCompany_Backend.Models.MailSenderModel;
+
+public class MailConfig
+{
+    public string Host { get; set; }
+    public int Port { get; set; }
+    public string SenderName { get; set; }
+    public string SenderMail { get; set; }
+    public string SenderPassword { get; set; }
+}
