+++ Models/MailSenderModels/MailContent.cs
@@ -1,0 +1,7 @@
+﻿namespace LethalCompany_Backend.Models.MailSenderModel;
+
+public class MailContent
+{
+    public string Subject { get; set; }
+    public string Body { get; set; }
+}
