+++ Models/TokenModels/GenerateTokenRequest.cs
@@ -1,0 +1,10 @@
+﻿using CommunifyLibrary.Models;
+
+namespace LethalCompany_Backend.Models.TokenModels;
+
+public class GenerateTokenRequest
+{
+    public string UserID { get; set; }
+    public Role Role { get; set; }
+    public DateTime ExpireDate { get; set; }
+}
