+++ Services/Interfaces/ITokenService.cs
@@ -1,6 +1,6 @@
 ﻿using CommunifyLibrary.Models;
+using LethalCompany_Backend.Models.TokenModels;
 using System.Security.Claims;
-using static Communify_Backend.Models.TokenModels;
 
 namespace Communify_Backend.Services.Interfaces
 {
